make static and animated pie charts
